Updates get/check for observations to check if the observation
has changed, else do nothing. Checks in both the app and view. This
is a possible refactor target, as there is a duplication in the
checking.

This update allows the app to handle going from no observations to
open observations, from observations to no observations, from an observation
to another observation, or staying the same through both logging in
and submitting the form.